% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiDirCDcML.R
\name{BiDirCDcML}
\alias{BiDirCDcML}
\title{Bi-directional CDcML Method}
\usage{
BiDirCDcML(
  b_X,
  b_Y,
  se_X,
  se_Y,
  n_X,
  n_Y,
  sig.cutoff = 5e-08,
  num_pert = 100,
  random_start = 0,
  random.seed = 0
)
}
\arguments{
\item{b_X}{Vector of estimated effect sizes for trait 1.}

\item{b_Y}{Vector of estimated effect sizes for trait 2.}

\item{se_X}{Vector of standard errors of b_X.}

\item{se_Y}{Vector of standard errors of b_Y.}

\item{n_X}{Sample size for trait 1.}

\item{n_Y}{Sample size for trait 2.}

\item{sig.cutoff}{Significant cutoff to choose IV, defaul is 5e-8.}

\item{num_pert}{Number of perturbations, default is 100.}

\item{random_start}{Number of starting points for solving non-convex optimization
in cML method, defaul is 0, i.e. no random starting point.}

\item{random.seed}{Random seed, default is 0, i.e. no random seed is set.}
}
\value{
List of estimated correlation ratio (est) and
corresponding standard error (se) obtained with CDcML, for
both directions. "NoS" indicates no screening, "S" indicates with screening.
"DP" indicates results with data perturbation, otherwise without data perturbation.
}
\description{
Bi-directional CDcML Method
}
