% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiDirCDMethod.R
\name{BiDirCDMethod}
\alias{BiDirCDMethod}
\title{Bi-directional CD Methods}
\usage{
BiDirCDMethod(
  b_X,
  b_Y,
  se_X,
  se_Y,
  n_X,
  n_Y,
  sig.cutoff = 5e-08,
  random.seed = 0
)
}
\arguments{
\item{b_X}{Vector of estimated effect sizes for trait 1.}

\item{b_Y}{Vector of estimated effect sizes for trait 2.}

\item{se_X}{Vector of standard errors of b_X.}

\item{se_Y}{Vector of standard errors of b_Y.}

\item{n_X}{Sample size for trait 1.}

\item{n_Y}{Sample size for trait 2.}

\item{sig.cutoff}{Significant cutoff to choose IV, defaul is 5e-8.}

\item{random.seed}{Random seed, default is 0, i.e. no random seed is set.}
}
\value{
List of estimated correlation ratio (est) and
corresponding standard error (se) obtained with CD-Ratio and CD-Egger, for
both directions. "NoS" indicates no screening, "S" indicates with screening.
}
\description{
Bi-directional CD Methods
}
